moved button to code container
